minimum coveringa
This script creates the minimum covering of N numbers by K in which we want to contain at least one combination of at least C matches. In this case we got 19631 combinations of 5 out of 45 numbers so that when 5 numbers are chosen we have at least one combination in our covering that has 4 matches.